+++ tests/data/invalid_data.py
@@ -84,3 +84,8 @@ INVALID_OPS_SET = {
     }
   }
 }
+
+INVALID_FILTER = {
+  "invalid_agent_name_in_filter" : ["ajdjd%&", "tata"],
+  "invalid_filter_type" : 0
+}