+++ tests/test_get_context_state.py
@@ -14,7 +14,7 @@ class TestGetContextStateSuccess(unittest.TestCase):
   @classmethod
   def setUpClass(cls):
     cls.client = CraftAIClient(settings.CRAFT_CFG)
-    cls.agent_id = valid_data.VALID_ID  + "_" + settings.RUN_ID
+    cls.agent_id = valid_data.VALID_ID  + "_" + settings.RUN_ID[-4:]
 
   def setUp(self):
     self.client.delete_agent(self.agent_id)
@@ -51,7 +51,7 @@ class TestGetContextStateFailure(unittest.TestCase):
   @classmethod
   def setUpClass(cls):
     cls.client = CraftAIClient(settings.CRAFT_CFG)
-    cls.agent_id = valid_data.VALID_ID  + "_" + settings.RUN_ID
+    cls.agent_id = valid_data.VALID_ID  + "_" + settings.RUN_ID[-4:]
 
   def setUp(self):
     self.client.delete_agent(self.agent_id)