+++ tests/test_get_shared_agent_inspector_url.py
@@ -10,7 +10,7 @@ class TestGetSharedAgentsInspectorUrlSuccess(unittest.TestCase):
   @classmethod
   def setUpClass(cls):
     cls.client = CraftAIClient(settings.CRAFT_CFG)
-    cls.agent_id = valid_data.VALID_ID  + "_" + settings.RUN_ID
+    cls.agent_id = valid_data.VALID_ID  + "_" + settings.RUN_ID[-4:]
 
   def setUp(self):
     self.client.delete_agent(self.agent_id)